requirements.txt
google-genai
openai
python-dotenv
faiss-cpu
numpy
sentence-transformers
fastapi
uvicorn[standard]
pydantic
python-multipart
httpx